requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
art==6.4
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
camelot-py==1.0.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.3
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.110.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.7.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.6.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.73.1
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
h5py==3.12.1
holidays==0.63
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
img2pdf==0.6.1
importlib_resources==6.4.5
invoice2data==0.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
langchain==0.3.26
langchain-core==0.3.68
langchain-google-genai==2.1.6
langchain-text-splitters==0.3.8
langsmith==0.4.4
Levenshtein==0.27.1
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.45.0
numpy==1.26.4
ocrmypdf==16.10.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.18
packaging==23.2
pandas==2.2.3
patsy==1.0.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pi_heif==0.22.0
pikepdf==9.7.0
pillow==10.3.0
pip==24.0
plotly==5.24.1
pluggy==1.6.0
propcache==0.3.2
prophet==1.1.6
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyodbc==5.2.0
pyparsing==3.2.0
pypdf==5.5.0
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
stanio==0.5.1
starlette==0.36.3
statsmodels==0.14.4
streamlit==1.35.0
sympy==1.14.0
tabula-py==2.10.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.5.0
tldextract==5.1.3
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.27.1
waitress==3.0.2
watchdog==6.0.0
Werkzeug==3.1.3
wfastcgi==3.0.0
wheel==0.45.1
wrapt==1.17.0
xgboost==2.1.3
yarl==1.20.1
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0